+++ source/blender/gpu/vulkan/vk_buffer.hh
@@ -19,7 +19,7 @@ class VKContext;
  * Class for handing vulkan buffers (allocation/updating/binding).
  */
 class VKBuffer {
-  int64_t size_in_bytes_ = 0;
+  size_t size_in_bytes_ = 0;
   VkBuffer vk_buffer_ = VK_NULL_HANDLE;
   VmaAllocation allocation_ = VK_NULL_HANDLE;
   /* Pointer to the virtually mapped memory. */
@@ -31,7 +31,7 @@ class VKBuffer {
 
   /** Has this buffer been allocated? */
   bool is_allocated() const;
-  bool create(int64_t size,
+  bool create(size_t size,
               GPUUsageType usage,
               VkBufferUsageFlags buffer_usage,
               bool is_host_visible = true);
