+++ source/blender/gpu/vulkan/vk_pixel_buffer.cc
@@ -12,7 +12,7 @@
 
 namespace blender::gpu {
 
-VKPixelBuffer::VKPixelBuffer(int64_t size) : PixelBuffer(size)
+VKPixelBuffer::VKPixelBuffer(size_t size) : PixelBuffer(size)
 {
   buffer_.create(
       size, GPU_USAGE_STATIC, VK_BUFFER_USAGE_TRANSFER_SRC_BIT | VK_BUFFER_USAGE_TRANSFER_DST_BIT);
